+++ app/src/main/java/com/dean/floatbackground/FloatText.java
@@ -19,7 +19,7 @@ public class FloatText extends FloatObject {
     }
 
     @Override
-    public void drawFloatObject(Canvas canvas, int x, int y, Paint paint) {
+    public void drawFloatObject(Canvas canvas, float x, float y, Paint paint) {
         paint.setTextSize(65);
         canvas.drawText(text, x, y, paint);
     }
